+++ CHANGELOG.md
@@ -4,10 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+### Added
+
+- Adds four circle blueprint
+- Adds brush size control
+- Adds modulo attribute to circles
+- Adds loading by index
+- Adds modulo controls
+
 ## [1.0.0]
 ### Added
 
 - Imports painter settings on storage import
+- Adds blueprint module
 
 ### Changed